users controll and default users

+++ resources/comments/controller.js
@@ -5,7 +5,7 @@ const router = express.Router();
 const passport = require('../../passport');
 
 
-router.put('/:id', ValidatorCheckOffensiveWord, async(req, res, next) => {
+router.put('/:id', passport.authenticate('jwt', { session: false }), ValidatorCheckOffensiveWord, async(req, res, next) => {
     try {
         const id = req.params.id;
         const comment = req.body;
@@ -23,15 +23,21 @@ router.put('/:id', ValidatorCheckOffensiveWord, async(req, res, next) => {
     }
 });
 
-router.delete('/:id', async(req, res, next) => {
+//poblisher solo puede borrar comentarios de sus entradas, admin todos
+router.delete('/:id',passport.authenticate('jwt', { session: false }), async(req, res, next) => {
     try {
-        const id = req.params.id;
-        const result = await CommentService.deleteComment(id);
-        if (result !== null) {
-            res.status(200).json(result);
-        }else{
-            res.status(404).json({message: 'Recurso no encontrado'})
+        const role =  req.user.role;
+        console.log(role)
+        if(role === 'admin'){
+            const id = req.params.id;
+            const result = await CommentService.deleteComment(id);
+            if (result !== null) {
+                res.status(200).json(result);
+            }else{
+                res.status(404).json({message: 'Recurso no encontrado'})
+            }
         }
+        
     } catch (error) {
         console.log(err);
         res.status(500).send(err);
